mutable virtual registers; basic allocation algorithm skeleton

+++ core_lang/src/asm_generation/compiler_constants.rs
@@ -1,2 +1,1 @@
-
-pub(crate) const NUM_FREE_REGISTERS: usize = 48;
+pub(crate) const NUM_FREE_REGISTERS: u8 = 48;
